+++ lib/raif.rb
@@ -10,6 +10,7 @@ require "raif/llm_registry"
 require "raif/embedding_model_registry"
 require "raif/json_schema_builder"
 require "raif/migration_checker"
+require "raif/messages"
 
 require "faraday"
 require "event_stream_parser"